host/HID/requirements.txt
hidapi==0.9.0.post3
numpy==1.24.3
